+++ x/nft-transfer/types/packet.go
@@ -73,5 +73,5 @@ func (nftpd NonFungibleTokenPacketData) ValidateBasic() error {
 
 // GetBytes is a helper for serializing
 func (nftpd NonFungibleTokenPacketData) GetBytes() []byte {
-	return sdk.MustSortJSON(ModuleCdc.MustMarshalJSON(&nftpd))
+	return ModuleCdc.MustMarshalJSON(&nftpd)
 }