+++ x/tieredrewards/keeper/grpc_query_test.go
@@ -1,0 +1,27 @@
+package keeper_test
+
+import (
+	"github.com/crypto-org-chain/chain-main/v8/x/tieredrewards/types"
+
+	sdkmath "cosmossdk.io/math"
+)
+
+func (s *KeeperSuite) TestGRPCQueryParams() {
+	customParams := types.NewParams(
+		sdkmath.LegacyNewDecWithPrec(3, 2),
+	)
+	s.keeper.InitGenesis(s.ctx, &types.GenesisState{Params: customParams})
+
+	resp, err := s.queryClient.Params(s.ctx.Context(), &types.QueryParamsRequest{})
+	s.Require().NoError(err)
+	s.Require().True(customParams.TargetBaseRewardsRate.Equal(resp.Params.TargetBaseRewardsRate))
+}
+
+func (s *KeeperSuite) TestGRPCQueryParams_Default() {
+	defaultGenesis := types.DefaultGenesisState()
+	s.keeper.InitGenesis(s.ctx, defaultGenesis)
+
+	resp, err := s.queryClient.Params(s.ctx.Context(), &types.QueryParamsRequest{})
+	s.Require().NoError(err)
+	s.Require().True(resp.Params.TargetBaseRewardsRate.IsZero())
+}