+++ x/tieredrewards/keeper/msg_server.go
@@ -1,0 +1,40 @@
+package keeper
+
+import (
+	"context"
+
+	"github.com/crypto-org-chain/chain-main/v8/x/tieredrewards/types"
+
+	"cosmossdk.io/errors"
+
+	govtypes "github.com/cosmos/cosmos-sdk/x/gov/types"
+)
+
+var _ types.MsgServer = msgServer{}
+
+// msgServer is a wrapper of Keeper.
+type msgServer struct {
+	Keeper
+}
+
+// NewMsgServerImpl returns an implementation of the tieredrewards MsgServer interface.
+func NewMsgServerImpl(k Keeper) types.MsgServer {
+	return &msgServer{Keeper: k}
+}
+
+// UpdateParams updates the module parameters.
+func (ms msgServer) UpdateParams(ctx context.Context, msg *types.MsgUpdateParams) (*types.MsgUpdateParamsResponse, error) {
+	if ms.authority != msg.Authority {
+		return nil, errors.Wrapf(govtypes.ErrInvalidSigner, "invalid authority; expected %s, got %s", ms.authority, msg.Authority)
+	}
+
+	if err := msg.Params.Validate(); err != nil {
+		return nil, err
+	}
+
+	if err := ms.Params.Set(ctx, msg.Params); err != nil {
+		return nil, err
+	}
+
+	return &types.MsgUpdateParamsResponse{}, nil
+}