+++ x/tieredrewards/types/keys.go
@@ -1,0 +1,16 @@
+package types
+
+import "cosmossdk.io/collections"
+
+var ParamsKey = collections.NewPrefix(0)
+
+const (
+	// ModuleName defines the module name
+	ModuleName = "tieredrewards"
+
+	// StoreKey defines the primary module store key
+	StoreKey = ModuleName
+
+	// RewardsPoolName defines the base reward pool module account name
+	RewardsPoolName = "rewards_pool"
+)